test: expand coverage for new settings and generator hooks
