+++ src/App.tsx
@@ -10,15 +10,10 @@ import SettingsPresets from './components/SettingsPresets';
 import TranslationLoader from './components/TranslationLoader';
 import { useTranslation } from './i18n';
 import './styles/components.css';
-import type {
-  MessageValue,
-  Operation,
-  PaperSizeOptions,
-  Problem,
-  QuizResult,
-  Settings,
-} from './types';
+import type { Operation, PaperSizeOptions, QuizResult, Settings } from './types';
 import { setupWCAGEnforcement } from './utils/wcagEnforcement';
+import { useProblemGenerator } from './hooks/useProblemGenerator';
+import { useSettings } from './hooks/useSettings';
 
 const SpeedInsights = React.lazy(() =>
   import('@vercel/speed-insights/react').then(module => ({ default: module.SpeedInsights }))
@@ -27,59 +22,18 @@ const SpeedInsights = React.lazy(() =>
 function App(): React.JSX.Element {
   const { t, isLoading } = useTranslation();
 
-  const defaultSettings: Settings = {
-    operations: ['+', '-'],
-    numProblems: 20,
-    numRange: [1, 20],
-    resultRange: [0, 20],
-    numOperandsRange: [2, 3],
-    allowNegative: false,
-    showAnswers: false,
-    fontSize: 16,
-    lineSpacing: 12,
-    paperSize: 'a4',
-  };
+  const { settings, setSettings, saveSettings, validateSettings } = useSettings();
+  const {
+    problems,
+    error,
+    warning,
+    successMessage,
+    generateProblems,
+    setError,
+    setWarning,
+    setSuccessMessage,
+  } = useProblemGenerator(settings, isLoading, validateSettings);
 
-  // Load settings from localStorage or use defaults
-  const loadSettings = (): Settings => {
-    try {
-      const saved = localStorage.getItem('mathgenie-settings');
-      if (saved) {
-        const parsed = JSON.parse(saved);
-        // Validate parsed settings to prevent errors
-        return {
-          ...defaultSettings,
-          ...parsed,
-          operations: Array.isArray(parsed.operations)
-            ? parsed.operations
-            : defaultSettings.operations,
-          numRange:
-            Array.isArray(parsed.numRange) && parsed.numRange.length === 2
-              ? parsed.numRange
-              : defaultSettings.numRange,
-          resultRange:
-            Array.isArray(parsed.resultRange) && parsed.resultRange.length === 2
-              ? parsed.resultRange
-              : defaultSettings.resultRange,
-          numOperandsRange:
-            Array.isArray(parsed.numOperandsRange) && parsed.numOperandsRange.length === 2
-              ? parsed.numOperandsRange
-              : defaultSettings.numOperandsRange,
-        };
-      }
-    } catch (error) {
-      if (process.env.NODE_ENV === 'development') {
-        console.warn('Failed to load settings from localStorage:', error);
-      }
-    }
-    return defaultSettings;
-  };
-
-  const [settings, setSettings] = useState<Settings>(() => loadSettings());
-  const [problems, setProblems] = useState<Problem[]>([]);
-  const [error, setError] = useState<MessageValue>('');
-  const [warning, setWarning] = useState<MessageValue>('');
-  const [successMessage, setSuccessMessage] = useState<MessageValue>('');
   const [isQuizMode, setIsQuizMode] = useState<boolean>(false);
   const [quizResult, setQuizResult] = useState<QuizResult | null>(null);
   const [hasInitialGenerated, setHasInitialGenerated] = useState<boolean>(false);
@@ -88,229 +42,6 @@ function App(): React.JSX.Element {
     a4: 'a4',
     letter: 'letter',
     legal: 'legal',
-  };
-
-  // Save settings to localStorage
-  const saveSettings = (newSettings: Settings): void => {
-    try {
-      localStorage.setItem('mathgenie-settings', JSON.stringify(newSettings));
-    } catch (error) {
-      if (process.env.NODE_ENV === 'development') {
-        console.warn('Failed to save settings to localStorage:', error);
-      }
-    }
-  };
-
-  // Validate settings
-  const validateSettings = (newSettings: Settings): MessageValue => {
-    if (newSettings.operations.length === 0) {
-      return { key: 'errors.noOperations' };
-    }
-    if (newSettings.numProblems <= 0 || newSettings.numProblems > 100) {
-      return { key: 'errors.invalidProblemCount' };
-    }
-    if (newSettings.numRange[0] > newSettings.numRange[1]) {
-      return { key: 'errors.invalidNumberRange' };
-    }
-    if (newSettings.resultRange[0] > newSettings.resultRange[1]) {
-      return { key: 'errors.invalidResultRange' };
-    }
-    if (
-      newSettings.numOperandsRange[0] > newSettings.numOperandsRange[1] ||
-      newSettings.numOperandsRange[0] < 2
-    ) {
-      return { key: 'errors.invalidOperandsRange' };
-    }
-    return '';
-  };
-
-  const calculateExpression = (operands: number[], operators: string[]): number | null => {
-    let result = operands[0];
-
-    for (let i = 0; i < operators.length; i++) {
-      const operator = operators[i];
-      const operand = operands[i + 1];
-
-      switch (operator) {
-        case '+':
-          result += operand;
-          break;
-        case '-':
-          result -= operand;
-          break;
-        case '*':
-          result *= operand;
-          break;
-        case '/':
-          if (operand === 0 || result % operand !== 0) {
-            return null;
-          }
-          result = result / operand;
-          break;
-        default:
-          return null;
-      }
-    }
-
-    return result;
-  };
-
-  const generateProblem = (): string => {
-    const crypto = window.crypto || (window as any).msCrypto;
-    const random = (min: number, max: number): number => {
-      const array = new Uint32Array(1);
-      crypto.getRandomValues(array);
-      const val = array[0] / (0xffffffff + 1);
-      return min + Math.floor((max - min + 1) * val);
-    };
-
-    const randomNonZero = (min: number, max: number): number | null => {
-      const values: number[] = [];
-      for (let i = min; i <= max; i++) {
-        if (i !== 0) {
-          values.push(i);
-        }
-      }
-      if (values.length === 0) {
-        return null;
-      }
-      const idx = random(0, values.length - 1);
-      return values[idx];
-    };
-
-    const buildExpression = (count: number): { operands: number[]; operators: string[] } | null => {
-      const operands = [random(settings.numRange[0], settings.numRange[1])];
-      const operators: string[] = [];
-      for (let i = 0; i < count - 1; i++) {
-        const operator = settings.operations[random(0, settings.operations.length - 1)];
-        operators.push(operator);
-        const next =
-          operator === '/'
-            ? randomNonZero(settings.numRange[0], settings.numRange[1])
-            : random(settings.numRange[0], settings.numRange[1]);
-        if (next === null) {
-          return null;
-        }
-        operands.push(next);
-      }
-      return { operands, operators };
-    };
-
-    const numOperands = random(settings.numOperandsRange[0], settings.numOperandsRange[1]);
-    if (numOperands < 2) {
-      return '';
-    }
-
-    const MAX_ATTEMPTS = 10000;
-    for (let attempt = 0; attempt < MAX_ATTEMPTS; attempt++) {
-      const expression = buildExpression(numOperands);
-      if (!expression) {
-        return '';
-      }
-      const { operands, operators } = expression;
-
-      const result = calculateExpression(operands, operators);
-      const invalid =
-        result === null ||
-        result < settings.resultRange[0] ||
-        result > settings.resultRange[1] ||
-        (!settings.allowNegative && result < 0);
-
-      if (invalid) {
-        continue;
-      }
-
-      let problem = operands[0].toString();
-      operators.forEach((operator, index) => {
-        problem += ` ${operator} ${operands[index + 1]}`;
-      });
-
-      problem = problem.replace(/\*/g, '✖').replace(/\//g, '➗');
-
-      return settings.showAnswers ? `${problem} = ${result}` : `${problem} = `;
-    }
-
-    return '';
-  };
-
-  const generateProblems = (showSuccessMessage: boolean = true): void => {
-    // Clear previous messages
-    setError('');
-    setWarning('');
-    setSuccessMessage('');
-
-    // Don't show any messages if i18n is still loading
-    if (isLoading) {
-      return;
-    }
-
-    // Validate settings first
-    const validationError = validateSettings(settings);
-    if (validationError) {
-      setError(validationError);
-      return;
-    }
-
-    // Show warning for large number of problems
-    if (settings.numProblems > 50) {
-      setWarning({
-        key: 'warnings.largeNumberOfProblems',
-        params: { count: settings.numProblems },
-      });
-    }
-
-    try {
-      const generatedProblems = Array.from({ length: settings.numProblems }, () =>
-        generateProblem()
-      )
-        .filter(problem => problem !== '')
-        .map((problem, index) => ({ id: index, text: problem }));
-
-      if (generatedProblems.length === 0) {
-        // Only show error if i18n is loaded and this is not initial generation
-        if (!isLoading && showSuccessMessage) {
-          setError({ key: 'errors.noProblemsGenerated' });
-        }
-      } else if (generatedProblems.length < settings.numProblems) {
-        // Partial generation - show warning only if i18n is loaded
-        if (!isLoading && showSuccessMessage) {
-          setWarning({
-            key: 'errors.partialGeneration',
-            params: {
-              generated: generatedProblems.length,
-              requested: settings.numProblems,
-            },
-          });
-        }
-        setProblems(generatedProblems);
-      } else {
-        // Full success - only show message if not initial load and i18n is loaded
-        if (showSuccessMessage && !isLoading) {
-          if (process.env.NODE_ENV === 'development') {
-            console.log('Setting success message:', {
-              showSuccessMessage,
-              isLoading,
-              count: generatedProblems.length,
-            });
-          }
-          setSuccessMessage({
-            key: 'messages.success.problemsGenerated',
-            params: { count: generatedProblems.length },
-          });
-          // Clear success message after 5 seconds
-          setTimeout(() => setSuccessMessage(''), 5000);
-        }
-        setProblems(generatedProblems);
-      }
-    } catch (err) {
-      // Only show error if i18n is loaded and this is not initial generation
-      if (!isLoading && showSuccessMessage) {
-        setError({ key: 'errors.generationFailed' });
-      }
-      if (process.env.NODE_ENV === 'development') {
-        console.error('Problem generation error:', err);
-      }
-    }
   };
 
   const downloadPdf = (): void => {
